+++ operator/tests/e2e/webhook_limits_test.go
@@ -173,7 +173,7 @@ func TestWebhookDeduplication(t *testing.T) {
 
 	// Wait for report.
 	reportName := expectedReportName(saName)
-	report := waitForPolicyReportCondition(ctx, t, reportName, ns, func(r *audiciav1alpha1.AudiciaPolicyReport) bool {
+	report := waitForPolicyReportCondition(ctx, t, reportName, ns, func(r *audiciav1alpha1.AudiciaReport) bool {
 		return r.Status.EventsProcessed > 0
 	}, defaultTimeout)